Refactors the linked list functions in ft_lst
