+++ libft/ft_lst/ft_lstadd_front.c
@@ -1,28 +1,31 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstadd_back.c                                   :+:      :+:    :+:   */
+/*   ft_lstadd_front.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vgoncalv <vgoncalv@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/05/23 14:44:58 by vgoncalv          #+#    #+#             */
-/*   Updated: 2021/05/23 14:44:58 by vgoncalv         ###   ########.fr       */
+/*   Created: 2021/05/23 14:28:42 by vgoncalv          #+#    #+#             */
+/*   Updated: 2022/09/05 16:19:38 by vgoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <stddef.h>
+#include <libft/ft_lst.h>
 
-void	ft_lstadd_back(t_list **lst, t_list *new)
+/**
+ * @brief Adds a `t_list` node to the start of the list.
+ * @param lst A pointer to the linked list
+ * @param new_node The new node
+ */
+void	ft_lstadd_front(t_list **lst, t_list *new_node)
 {
-	t_list	*temp;
-
-	if (lst == NULL || new == NULL)
+	if (lst == NULL || new_node == NULL)
 		return ;
-	if (*lst == NULL)
-		*lst = new;
-	else
+	if (*lst != NULL)
 	{
-		temp = ft_lstlast(*lst);
-		temp->next = new;
+		new_node->next = *lst;
+		(*lst)->prev = new_node;
 	}
+	*lst = new_node;
 }
