+++ libft/ft_lst/ft_lstclear.c
@@ -1,26 +1,33 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstsize.c                                       :+:      :+:    :+:   */
+/*   ft_lstclear.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vgoncalv <vgoncalv@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/05/23 14:33:30 by vgoncalv          #+#    #+#             */
-/*   Updated: 2021/05/23 14:33:30 by vgoncalv         ###   ########.fr       */
+/*   Created: 2021/05/23 15:02:44 by vgoncalv          #+#    #+#             */
+/*   Updated: 2022/09/05 16:05:05 by vgoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <stddef.h>
+#include <libft/ft_lst.h>
 
-int	ft_lstsize(t_list *lst)
+/**
+ * @brief Frees the allocated memory of all the nodes in a linked list.
+ * @param lst The linked list
+ * @param del The content deletion function
+ */
+void	ft_lstclear(t_list **lst, void (*del)(void *))
 {
-	int	size;
+	t_list	*temp;
 
-	size = 0;
-	while (lst != NULL)
+	if (lst == NULL || *lst == NULL)
+		return ;
+	while (*lst != NULL)
 	{
-		lst = lst->next;
-		size++;
+		temp = (*lst)->next;
+		ft_lstdelone(*lst, del);
+		*lst = temp;
 	}
-	return (size);
 }
