+++ libft/ft_lst/ft_lstdelone.c
@@ -1,22 +1,28 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstadd_front.c                                  :+:      :+:    :+:   */
+/*   ft_lstdelone.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vgoncalv <vgoncalv@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/05/23 14:28:42 by vgoncalv          #+#    #+#             */
-/*   Updated: 2021/05/23 14:28:42 by vgoncalv         ###   ########.fr       */
+/*   Created: 2021/05/23 14:55:12 by vgoncalv          #+#    #+#             */
+/*   Updated: 2022/09/05 16:05:18 by vgoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <stdlib.h>
+#include <libft/ft_lst.h>
 
-void	ft_lstadd_front(t_list **lst, t_list *new)
+/**
+ * @brief Frees the allocated memory of a nodes in a linked list.
+ * @param lst The linked list
+ * @param del The content deletion function
+ */
+void	ft_lstdelone(t_list *lst, void (*del)(void *))
 {
-	if (lst == NULL)
-		return ;
-	if (*lst != NULL)
-		new->next = *lst;
-	*lst = new;
+	if (lst != NULL)
+	{
+		del(lst->content);
+		free(lst);
+	}
 }
