+++ libft/ft_lst/ft_lstlast.c
@@ -1,22 +1,27 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstiter.c                                       :+:      :+:    :+:   */
+/*   ft_lstlast.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vgoncalv <vgoncalv@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/05/23 15:20:55 by vgoncalv          #+#    #+#             */
-/*   Updated: 2021/05/23 15:20:55 by vgoncalv         ###   ########.fr       */
+/*   Created: 2021/05/23 14:35:32 by vgoncalv          #+#    #+#             */
+/*   Updated: 2022/09/05 16:09:41 by vgoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <stddef.h>
+#include <libft/ft_lst.h>
 
-void	ft_lstiter(t_list *lst, void (*f)(void *))
+/**
+ * @brief Returns the last node of a linked list.
+ * @param lst The first node of the list
+ */
+t_list	*ft_lstlast(t_list *lst)
 {
-	while (lst != NULL)
-	{
-		f(lst->content);
+	if (lst == NULL)
+		return (NULL);
+	while (lst->next != NULL)
 		lst = lst->next;
-	}
+	return (lst);
 }
