+++ libft/ft_lst/ft_lstnew.c
@@ -6,13 +6,19 @@
 /*   By: vgoncalv <vgoncalv@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/23 14:26:27 by vgoncalv          #+#    #+#             */
-/*   Updated: 2022/09/04 16:19:03 by vgoncalv         ###   ########.fr       */
+/*   Updated: 2022/09/05 16:00:44 by vgoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <stddef.h>
 #include <libft/ft_mem.h>
+#include <libft/ft_lst.h>
 
+/**
+ * @brief Creates a new `t_list` node.
+ * @param content A pointer to the content of the node
+ * @return The allocated  `t_list` node
+ */
 t_list	*ft_lstnew(void *content)
 {
 	t_list	*item;
