--- libft/ft_lstdelone.c
@@ -1,22 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_lstdelone.c                                     :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: vgoncalv <vgoncalv@student.42sp.org.br>    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/05/23 14:55:12 by vgoncalv          #+#    #+#             */
-/*   Updated: 2021/05/23 14:55:12 by vgoncalv         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-void	ft_lstdelone(t_list *lst, void (*del)(void *))
-{
-	if (lst != NULL)
-	{
-		del(lst->content);
-		free(lst);
-	}
-}
